add tables to display qualities in the character sheet
heaven help me I vibe coded this, but it turned out ok. Might be an ok to work on this a bit to give to an agent to keep extending this app. But I know that some functionality is too complex for most agent models. Also I did have to prompt it a few times to get an output that I was more comfortable with.